requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
anyio==4.8.0
bitsandbytes==0.45.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.6
filelock==3.17.0
fsspec==2024.12.0
h11==0.14.0
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
passlib==1.7.4
psutil==6.1.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
